Transferring functionality from ActivityA to ActivityC
