+++ app/src/main/res/layout/activity_a.xml
@@ -142,27 +142,6 @@
         android:hint="@string/et_password_hint"
         android:inputType="textPassword" />
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_marginTop="@dimen/middle_margin">
-
-        <TextView
-            android:id="@+id/tv_bench_press"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/tv_bench_press"
-            android:textSize="@dimen/header2_size" />
-
-        <Spinner
-            android:id="@+id/sp_bench_press"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_marginStart="10dp" />
-
-    </LinearLayout>
-
     <Button
         android:id="@+id/btn_toast"
         android:layout_width="wrap_content"
