+++ app/src/main/res/layout/activity_c.xml
@@ -202,5 +202,23 @@
         android:orientation="vertical"
         app:layout_constraintGuide_begin="@dimen/right_guideline" />
 
+    <TextView
+        android:id="@+id/tv_bench_press"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="@string/tv_bench_press"
+        android:textSize="@dimen/header2_size"
+        android:layout_marginTop="@dimen/middle_margin"
+        app:layout_constraintTop_toBottomOf="@id/cb_football"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <Spinner
+        android:id="@+id/sp_bench_press"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginStart="10dp"
+        app:layout_constraintTop_toTopOf="@id/tv_bench_press"
+        app:layout_constraintBottom_toBottomOf="@id/tv_bench_press"
+        app:layout_constraintStart_toEndOf="@id/tv_bench_press"/>
 
 </androidx.constraintlayout.widget.ConstraintLayout>
